Add fullscreen mode, inventory textures, and scroll bar
- Changed window creation mode to fullscreen in `Globals.cpp`.
- Added `InventoryTextures_PG.h` and `ScrollBar_PG.h` includes.
- Replaced `std::unique_ptr` with direct instances in texture handlers.
- Introduced new textures for inventory and scroll bars.
- Added methods for inventory display and scroll bar handling.
- Implemented `addButton` method in `Menu_PG.h`.
- Added static casts for type safety in `AstarAlgorithm_PG.cpp`.
- Removed unnecessary offset in `EnemyShipHandler_PG.cpp`.
- Updated `InventoryMenu_PG.cpp` for new textures and scaling.
- Updated README with troubleshooting and setup steps.
- Added new source files and headers to project files.
- Corrected typo in `water_tiles_sheet.png` in project file.
- Organized source files and sprites in project filters.
- Added new PNG and Aseprite image files for UI elements.
- Implemented `ScrollBar` class in `ScrollBar_PG.cpp`.

+++ PirateGame/Resource/AstarAlgorithm_PG.cpp
@@ -136,7 +136,7 @@ std::vector<sf::Vector2i> AStar::getNeighbors(const sf::Vector2i& node) {
 }
 
 float AStar::heuristic(const sf::Vector2i& a, const sf::Vector2i& b) const {
-    return std::sqrt((a.x - b.x) * (a.x - b.x) + (a.y - b.y) * (a.y - b.y));
+    return static_cast<float>(std::sqrt((a.x - b.x) * (a.x - b.x) + (a.y - b.y) * (a.y - b.y)));
 }
 
 std::vector<sf::Vector2i> AStar::reconstructPath(std::shared_ptr<AstarNode> node) const {
@@ -150,11 +150,11 @@ std::vector<sf::Vector2i> AStar::reconstructPath(std::shared_ptr<AstarNode> node
 }
 
 sf::Vector2i AStar::vectorToGrid(const sf::Vector2f& vector) const {
-    return sf::Vector2i(std::floor(vector.x / tileSize), std::floor(vector.y / tileSize));
+    return sf::Vector2i(std::floor(vector.x / tileSize), std::floor(vector.y / static_cast<float>(tileSize)));
 }
 
 sf::Vector2f AStar::gridToVector(const sf::Vector2i& grid) const {
-    return sf::Vector2f(static_cast<int>(grid.x) * tileSize, static_cast<int>(grid.y) * tileSize);
+    return sf::Vector2f(grid.x * tileSize, grid.y * tileSize);
 }
 
 bool AStar::isObstacle(const sf::Vector2i& node) const {
